+++ app/src/main/java/me/ccrama/redditslide/SubmissionViews/PopulateShadowboxInfo.kt
@@ -875,7 +875,7 @@ object PopulateShadowboxInfo {
             withContext(Dispatchers.IO) {
                 postRepository.likePost(
                     (submission as LemmyPost).instance,
-                    PostId(submission.rowId),
+                    PostId(submission.rowId.toInt()),
                     direction
                 )
             }
